% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explain_One_Way_ANOVA.R
\name{SLExplain.ANOVA_one_way}
\alias{SLExplain.ANOVA_one_way}
\title{Explain One-way ANOVA}
\usage{
SLExplain.ANOVA_one_way()
}
\value{
Invisibly returns NULL.
}
\description{
Prints an explanation of One-way ANOVA, it's definition,
interpretation, examples and setup
}
