% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explain_Chi_Square.R
\name{SLExplain.chi_square}
\alias{SLExplain.chi_square}
\title{Explain Chi-Square test}
\usage{
SLExplain.chi_square()
}
\value{
Invisibly returns NULL.
}
\description{
Prints an explanation of Chi-Square test, it's definition,
interpretation, examples and setup
}
