% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explain_Chi_Square_Goodness_of_Fit.R
\name{SLExplain.chi_square_goodness}
\alias{SLExplain.chi_square_goodness}
\title{Explain Chi-Square Goodness of Fit}
\usage{
SLExplain.chi_square_goodness()
}
\value{
Invisibly returns NULL.
}
\description{
Prints an explanation of Chi-Square Goodness of Fit, its definition,
interpretation, examples and setup
}
