% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explain_levene.R
\name{SLExplain.levene}
\alias{SLExplain.levene}
\title{Explain Levenes test}
\usage{
SLExplain.levene()
}
\value{
Invisibly returns NULL.
}
\description{
Prints and explenation of Levenes, it's definition,
interpretation, examples and setup
}
