% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explain_McNemar.R
\name{SLExplain.mcnemar}
\alias{SLExplain.mcnemar}
\title{Explain McNemar's test}
\usage{
SLExplain.mcnemar()
}
\value{
Invisibly returns NULL.
}
\description{
Prints an explanation of McNemar's test, it's definition,
interpretation, examples and setup
}
