% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explain_p_value.R
\name{SLExplain.p_value}
\alias{SLExplain.p_value}
\title{Explain p-values}
\usage{
SLExplain.p_value()
}
\value{
Invisibly returns NULL.
}
\description{
Prints and explenation of p-values, it's definition, interpretation, 
some misconceptions, and connection to hypothesis testing
}
