% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explain_shapiro_wilks.R
\name{SLExplain.shapirowilks}
\alias{SLExplain.shapirowilks}
\title{Explain Shapiro Wilks Normality test}
\usage{
SLExplain.shapirowilks()
}
\value{
Invisibly returns NULL.
}
\description{
Prints and explenation of Shapiro Wilks, it's definition,
interpretation, examples and setup
}
