% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explain_Significance_Levels.R
\name{SLExplain.significance_level}
\alias{SLExplain.significance_level}
\title{Explain significance levels}
\usage{
SLExplain.significance_level()
}
\value{
Invisibly returns NULL.
}
\description{
Prints and explenation of significance levels, it's definition,
interpretation, some misconceptions, and connection to hypothesis testing
}
