% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explain_t_test_one_sample.R
\name{SLExplain.t_test_one_sample}
\alias{SLExplain.t_test_one_sample}
\title{Explain t-test}
\usage{
SLExplain.t_test_one_sample()
}
\value{
Invisibly returns NULL.
}
\description{
Prints and explenation of t-tests, it's definition,
interpretation, some misconceptions, and connection to hypothesis testing
Includes assumptions, examples and setup
}
