% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explain_t_test_paired.R
\name{SLExplain.t_test_paired}
\alias{SLExplain.t_test_paired}
\title{Explain Paired sample t-test}
\usage{
SLExplain.t_test_paired()
}
\value{
Invisibly returns NULL.
}
\description{
Prints and explenation of Paired sample t-test, it's definition,
interpretation, examples and setup
}
