% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explain_t_test_pairwise.R
\name{SLExplain.t_test_pairwise}
\alias{SLExplain.t_test_pairwise}
\title{Explain Pairwise t-test (Multiple paired t-test)}
\usage{
SLExplain.t_test_pairwise()
}
\value{
Invisibly returns NULL.
}
\description{
Prints an explanation of Pairwise t-test (Multiple paired t-test),
it's definition, interpretation, examples and setup
}
