% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explain_t_test_two_sample.R
\name{SLExplain.t_test_two_sample}
\alias{SLExplain.t_test_two_sample}
\title{Explain Independent two-sample t-test}
\usage{
SLExplain.t_test_two_sample()
}
\value{
Invisibly returns NULL.
}
\description{
Prints and explenation of two-sample t-test, it's definition,
interpretation, examples and setup
}
