% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explain_Tukeys_test.R
\name{SLExplain.tukeys}
\alias{SLExplain.tukeys}
\title{Explain Tukey's test}
\usage{
SLExplain.tukeys()
}
\value{
Invisibly returns NULL.
}
\description{
Prints an explanation of Tukey's test, it's definition,
interpretation, examples and setup
}
